requirements.txt
PyExecJS==1.5.1
pyperclip==1.8.2
requests==2.26.0
